Text/CustomNER/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

jsonschema
azure-functions
azure-core
azure-ai-textanalytics==5.2.0b4